refactor: Move User and UserRole interfaces to a separate file for better organization
